REFACTORING added support for batch uploads
Had to rework since we're allowing batch uploads. Before, I was writing
the contents of the request into memory which wasn't immediately obvious
as an issue... whoops. But if the client tries uploading multiple videos
around ~500mb in size, then this would be super problematic and the
server would just run OOM.

Also some changes to naming.